Standardize weight on kilograms everywhere
The DB column was already Float/kg (with a weight_grams_to_kg data migration),
but the app still treated weight as integer grams. Aligned the whole stack to kg:

- Schemas: child.weight / childVisit.weight / health.weight are now z.number()
  (decimals allowed) instead of z.number().int(); comments say kilograms.
- Backend: removed the /1000 grams→kg conversions in health route and
  child.service z-score calc — weight is already kg.
- Frontend: removed /1000 display conversions (Dashboard, ChildDetail,
  ChildVisitDetail, FamilyDetail); relabeled inputs "Weight (kg)" with step=0.1.
- Tests: converted all gram-scale weight fixtures/assertions to kg.

No schema migration needed (column already Float). Existing gram data is
converted by the existing weight_grams_to_kg migration — see caveat re:
whether it has already run on a given database.

599 tests pass; build type-checks.

+++ packages/shared/src/schemas/child-visit.ts
@@ -12,7 +12,7 @@ export const ChildVisitCreateSchema = z.object({
   familyId: z.number().int().positive(),
   childId: z.number().int().positive(),
   visitDate: z.string().datetime(),
-  weight: z.number().int().min(0).default(0), // grams
+  weight: z.number().min(0).default(0), // kilograms
   armCircumference: z.number().int().min(0).default(0), // millimeters
   height: z.number().int().min(0).default(0), // millimeters
   incap: z.boolean().default(false), // gave special drink
@@ -37,7 +37,7 @@ export const ChildVisitReadSchema = z.object({
   familyId: z.number().int().positive(),
   childId: z.number().int().positive(),
   visitDate: z.string().datetime(),
-  weight: z.number().int(),
+  weight: z.number(),
   armCircumference: z.number().int(),
   height: z.number().int(),
   incap: z.boolean(),
